LayoutTests/inspector/model/remote-object-weak-collection-expected.txt


-----------------------------------------------------
EXPRESSION:  weakMap
ENTRIES:
[]

-----------------------------------------------------
EXPRESSION: weakMap.set(strongKey1, 1); weakMap.set(strongKey2, 2); weakMap
ENTRIES:
[
  {
    "key": {
      "type": "object",
      "objectId": "<filtered>",
      "className": "Object",
      "description": "Object",
      "preview": {
        "type": "object",
        "description": "Object",
        "lossless": true,
        "properties": [
          {
            "name": "id",
            "type": "number",
            "value": "1"
          }
        ]
      }
    },
    "value": {
      "type": "number",
      "value": 1,
      "description": "1"
    }
  },
  {
    "key": {
      "type": "object",
      "objectId": "<filtered>",
      "className": "Object",
      "description": "Object",
      "preview": {
        "type": "object",
        "description": "Object",
        "lossless": true,
        "properties": [
          {
            "name": "id",
            "type": "number",
            "value": "2"
          }
        ]
      }
    },
    "value": {
      "type": "number",
      "value": 2,
      "description": "2"
    }
  }
]

-----------------------------------------------------
EXPRESSION: delete window.strongKey1; weakMap
ENTRIES:
[
  {
    "key": {
      "type": "object",
      "objectId": "<filtered>",
      "className": "Object",
      "description": "Object",
      "preview": {
        "type": "object",
        "description": "Object",
        "lossless": true,
        "properties": [
          {
            "name": "id",
            "type": "number",
            "value": "2"
          }
        ]
      }
    },
    "value": {
      "type": "number",
      "value": 2,
      "description": "2"
    }
  }
]

-----------------------------------------------------
EXPRESSION: weakMap.set({id:3}, 3); weakMap.set({id:4}, 4); weakMap
ENTRIES:
[
  {
    "key": {
      "type": "object",
      "objectId": "<filtered>",
      "className": "Object",
      "description": "Object",
      "preview": {
        "type": "object",
        "description": "Object",
        "lossless": true,
        "properties": [
          {
            "name": "id",
            "type": "number",
            "value": "2"
          }
        ]
      }
    },
    "value": {
      "type": "number",
      "value": 2,
      "description": "2"
    }
  }
]

-----------------------------------------------------
EXPRESSION: delete window.strongKey2; weakMap
ENTRIES:
[]
